Version 5.0

+++ views/mascota/lista.php
@@ -27,11 +27,12 @@
     			   echo "<td>$mascota->sexo</td>";
     			   echo "<td>$mascota->biografia</td>";
     			   echo "<td>$mascota->fechanacimiento</td>";
-    			   echo "<td>$mascota->fechafallecimiento</td>";
-    			   echo "<td>";
+    			   echo "<td>";    			   
     			   echo " <a href='/mascota/show/$mascota->id'>Ver</a>";
+    			   if(Login::isAdmin() || Login::hasPrivilege(500)){
     			   echo "-<a href='/mascota/edit/$mascota->id'>Actualizar</a>";
     			   echo "-<a href='/mascota/delete/$mascota->id'>Borrar</a>";
+    			   }
     			   echo "</td>";
     			   echo "</tr>";
     		}?>
